remove queue resolver
